+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # BOCD.jl changelog
 
+## Unreleased
+### Added
+- Gaussian unknown mean with known variance conjugate model
+
+### Changed
+- README.md with new instructions
+- minor typo errors
+
 ## Version 1
 
 ### Version 0.1.0